+++ src/dark_energy.rs
@@ -1,7 +1,7 @@
 use crate::Redshift;
 
 pub trait DarkEnergyEquationOfState {
-    fn w(&self, z: Vec<Redshift>) -> Vec<f32>;
+    fn w(&self, z: Vec<Redshift>) -> Vec<f64>;
 }
 
 // TODO: impl DarkEnergyEquationOfState for Cosmology
